+++ provider/internal/device/ftd/resource.go
@@ -35,41 +35,6 @@ type Resource struct {
 	client *cdoClient.Client
 }
 
-var _ resource.ConfigValidator = &performanceTierConfigValidator{}
-
-type performanceTierConfigValidator struct{}
-
-func (c performanceTierConfigValidator) Description(ctx context.Context) string {
-	return c.MarkdownDescription(ctx)
-}
-
-func (c performanceTierConfigValidator) MarkdownDescription(ctx context.Context) string {
-	return fmt.Sprintf("TODO")
-}
-
-func (c performanceTierConfigValidator) ValidateResource(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
-	var configData ResourceModel
-	resp.Diagnostics.Append(req.Config.Get(ctx, &configData)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// if configData virtual is set to true
-	if !configData.Virtual.IsUnknown() && !configData.Virtual.IsNull() && configData.Virtual.ValueBool() {
-		// and performance tier is not set
-		if configData.PerformanceTier.IsNull() || configData.PerformanceTier.IsNull() {
-			// then we error
-			resp.Diagnostics.AddError("Performance Tier is required for virtual FTD.", "You need to select a performance tiers for virtual FTD. Allowed values are: [\"FTDv5\", \"FTDv10\", \"FTDv20\", \"FTDv30\", \"FTDv50\", \"FTDv100\", \"FTDv\"].")
-		}
-	}
-}
-
-func (r *Resource) ConfigValidators(ctx context.Context) []resource.ConfigValidator {
-	return []resource.ConfigValidator{
-		performanceTierConfigValidator{},
-	}
-}
-
 type ResourceModel struct {
 	ID               types.String `tfsdk:"id"`
 	Name             types.String `tfsdk:"name"`
@@ -298,4 +263,10 @@ func (r *Resource) Delete(ctx context.Context, req resource.DeleteRequest, res *
 
 func (r *Resource) ImportState(ctx context.Context, req resource.ImportStateRequest, res *resource.ImportStateResponse) {
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, res)
+}
+
+func (r *Resource) ConfigValidators(ctx context.Context) []resource.ConfigValidator {
+	return []resource.ConfigValidator{
+		performanceTierConfigValidator{},
+	}
 }
